reporte de km

+++ vehiculos-service/src/main/java/org/backend/Repository/VehiculoRepository.java
@@ -4,6 +4,6 @@ import org.backend.entities.Vehiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
 
 }
